+++ FCNameColor/API/FCNameColorProvider.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using Dalamud.Logging;
 using Dalamud.Plugin;
 using Dalamud.Plugin.Ipc;
 using Dalamud.Plugin.Services;
